requirements.txt
alembic           
annotated-types   
anyio             
asyncpg            
bcrypt              
cffi              
click               
colorama          
cryptography      
ecdsa             
fastapi           
greenlet          
h11               
idna              
Jinja2           
Mako              
MarkupSafe        
passlib           
pip               
pyasn1            
pycparser         
pydantic          
pydantic_core     
pyOpenSSL         
python-jose       
python-multipart  
python-slugify    
rsa               
six              
sniffio           
SQLAlchemy       
starlette         
text-unidecode    
typing_extensions 
uvicorn           